Flow/202601/2026-01-03/clinic_list_downloader/requirements.txt
# Clinic List Downloader - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Requests
requests==2.31.0

# Data Validation
pydantic==2.5.3

# ASGI Server
python-multipart==0.0.6

# Server-Sent Events
sse-starlette==1.8.2

# Cloud Storage (for database persistence)
google-cloud-storage==2.14.0
